requirements.txt
# AI PPT Generator Dependencies

# Core PPT manipulation
python-pptx==0.6.23

# Web API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic[email]==2.5.0

# AI
openai==1.3.0

# HTTP
requests==2.31.0

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.0
